# gh-128807: Add marking phase for free-threaded cyclic GC