--- a/Include/internal/pycore_gc.h
+++ b/Include/internal/pycore_gc.h
@@ -51,7 +51,7 @@ static inline PyObject* _Py_FROM_GC(PyGC_Head *gc) {
 #  define _PyGC_BITS_FROZEN         (1<<3)
 #  define _PyGC_BITS_SHARED         (1<<4)
 #  define _PyGC_BITS_ALIVE          (1<<5)    // Reachable from a known root.
-#  define _PyGC_BITS_DEFERRED       (1<<7)    // Use deferred reference counting
+#  define _PyGC_BITS_DEFERRED       (1<<6)    // Use deferred reference counting
 #endif
 
 #ifdef Py_GIL_DISABLED