This is conceptually similar to the phase that was added to the non-free-threaded GC.  Start with a set of known root objects, like sysdict and mark all objects reachable from those (revealed by the tp_traverse method) as "alive".  We know anything marked alive cannot be garbage and can be excluded from the regular cyclic GC process.  For most programs, this saves a moderate amount of computation since the marking pass is relatively cheaper per object.

If `gc.freeze()` is used, it's unlikely that this marking phase will be a win since it's expected that the majority of objects will be frozen.  Disable the marking phase if freeze is used.

See gh-126491 for the non-free-threaded version of this technique.

pyperformance [results vs merge base](https://github.com/facebookexperimental/free-threading-benchmarking/blob/main/results/bm-20250113-3.14.0a3%2B-0c173c9-NOGIL/bm-20250113-vultr-x86_64-nascheme-nogil_gc_mark_alive-3.14.0a3%2B-0c173c9-vs-base.svg).  I suspect the slowdown on some benchmarks is not real.  For example, regex_v8 should not be slower.

[Here](https://gist.github.com/nascheme/e18406f977b3d37ef3d9813320e8faa9) are the pyperformance results for a bare-metal AMD Ryzen machine.  It does not show a slowdown on regex_v8, for example.

To better show the expected improvement, I ran a "sphinx build" benchmark, like in gh-124567.  Results are:

|   | old  | new  |
|---|---|---|
| GC collections  | 38  | 38  | 
| long-lived objects  |  586,986  |  585,730  |
| total run time  | 3.24 s | 3.03 s  |
| mark phase time | 0.00 s | 0.16 s |
| total gc time | 0.72 s | 0.57 s |



<!-- gh-issue-number: gh-128807 -->
* Issue: gh-128807
<!-- /gh-issue-number -->